added components and routing features
