+++ src/index.css
@@ -29,6 +29,70 @@ h2 {
   text-shadow: black 2px 2px;
 }
 
+/*Splash Screen*/
+.splash-hero {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: url('./assets/lightning-in-the-sea.jpg') no-repeat center center/cover;
+}
+
+.splash-welcome {
+  font-size: 4rem;
+  margin-bottom: 2rem;
+  font-family: 'Oswald', sans-serif;
+}
+
+.splash-container {
+  width: 50%;
+  height: 50%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.splash-welcome {
+  color: rgb(400,400,400);
+}
+
+.splash-options {
+  display: flex;
+  list-style: none;
+  justify-content: space-between;
+}
+
+.splash-options a, .splash-options a:visited, .splash-options a:active {
+  color: white;
+  text-decoration: none;
+  font-family: 'Oswald', sans-serif;
+}
+
+.splash-search-tool {
+  background-color: rgba(255,255,255,0.2);
+  padding: 2rem;
+  margin-right: 1rem;
+  border-radius: 10px;
+}
+
+.splash-search-tool:hover {
+  background-color: rgba(200,200,200,0.8);
+  color: black;
+}
+
+.splash-about {
+  background-color: rgba(255,255,255,0.2);
+  padding: 2rem;
+  margin-left: 1rem;
+  border-radius: 10px;
+}
+.splash-about:hover {
+  background-color: rgba(200,200,200,0.8);
+  color: black;
+}
+
 /*weather-app screen properties */
 .weather-app {
   width: 100%;
@@ -53,6 +117,35 @@ h2 {
 .weather-app .search {
   text-align: center;
   padding: 1rem;
+  margin-bottom: 50px;
+}
+
+.back-button {
+  margin: 0 auto;
+  padding: .9rem 1.5rem;
+  width: 200px;
+  font-size: 1.4rem;
+  text-align: center;
+  border-radius: 25px;
+  border: 1px solid rgba(255, 255, 255, 0.8);
+  background: rgba(255, 255, 255, 0.1);
+  color: #f8f8f8;
+  -webkit-box-shadow: 0px 2px 5px 0px rgba(230, 230, 230, 1);
+  -moz-box-shadow: 0px 2px 5px 0px rgba(230, 230, 230, 1);
+  box-shadow: 0px 2px 5px 0px rgba(230, 230, 230, 1);
+  transition: 1s;
+}
+
+.back-button a {
+  text-decoration: none;
+  color: #f8f8f8;
+}
+
+.back-button:hover {
+  margin-top: 30px;
+  -webkit-box-shadow: 0px 2px 25px 0px rgba(230, 230, 230, 1);
+  -moz-box-shadow: 0px 2px 25px 0px rgba(230, 230, 230, 1);
+  box-shadow: 0px 2px 25px 0px rgba(230, 230, 230, 1);
 }
 
 .weather-app input {
@@ -127,7 +220,7 @@ h2 {
 /* about-us container */
 .container-about-us {
   max-width: 700px;
-  height: 500px;
+  height: 700px;
   margin: auto;
   padding: 0 1rem;
   position: relative;
